agents/pyproject.toml
[project]
name = "autonomx-agents"
version = "0.1.0"
description = "AutoNomX AI Agent Runtime"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "protobuf>=5.29.0",
    "litellm>=1.55.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "psycopg[binary]>=3.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
